Move calculation functions for PA1 to pa1 module
